refactor(detail): extract DetailInfoRail + DetailVersionsDialog from ForgeEntryDetail
ForgeEntryDetail had grown to a 450-line component mixing layout, the owner
islands, the Information sidebar and the versions popup. Pulled the two largest
self-contained server regions into their own components:

- DetailInfoRail: the persistent 'Information' rail (the metadata <dl>), with the
  isHttpUrl link guard that only it uses. Takes the resolved metadata primitives
  as props.
- DetailVersionsDialog: the native <dialog> versions popup (version views + the
  owner-only upload-new-version form), building its own ADD_VERSION_LABELS.

ForgeEntryDetail (450 -> 321 lines) now composes them. No behaviour change: every
data hook is preserved (data-detail-info, data-tag-list, data-versions-dialog,
data-add-version), the two-tablist structure and the AAA landmark layout are
unchanged. Verified against the redeployed module: detail/tabs/authoring/
prepackaged 32/32, accessibility 5/5 (detail page + editor + lightbox AAA),
storefront 22/22.
